requirements.txt
openai==1.3.0
requests==2.31.0
pyyaml==6.0.1
jsonschema==4.19.2
streamlit==1.28.1
jinja2==3.1.2
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
pydantic==2.5.0
typing-extensions==4.8.0 